Undo view info change, use explicit alias
