+++ src/parser/query_node/select_node.cpp
@@ -1,11 +1,67 @@
 #include "duckdb/parser/query_node/select_node.hpp"
 #include "duckdb/parser/expression_util.hpp"
+#include "duckdb/parser/expression/case_expression.hpp"
+#include "duckdb/parser/expression/subquery_expression.hpp"
+#include "duckdb/parser/parsed_expression_iterator.hpp"
 #include "duckdb/common/serializer/serializer.hpp"
 
 namespace duckdb {
 
 SelectNode::SelectNode()
     : QueryNode(QueryNodeType::SELECT_NODE), aggregate_handling(AggregateHandling::STANDARD_HANDLING) {
+}
+
+static bool ContainsSimpleCase(QueryNode &node);
+
+static bool ContainsSimpleCase(const ParsedExpression &expression) {
+	if (expression.GetExpressionClass() == ExpressionClass::CASE && expression.Cast<CaseExpression>().CaseOperand()) {
+		return true;
+	}
+	if (expression.GetExpressionClass() == ExpressionClass::SUBQUERY) {
+		auto &subquery = expression.Cast<SubqueryExpression>().Subquery();
+		if (subquery && ContainsSimpleCase(*subquery->node)) {
+			return true;
+		}
+	}
+	for (auto &child : expression.Children()) {
+		if (ContainsSimpleCase(child)) {
+			return true;
+		}
+	}
+	return false;
+}
+
+static bool ContainsSimpleCase(QueryNode &node) {
+	bool result = false;
+	ParsedExpressionIterator::EnumerateQueryNodeChildren(
+	    node, [&](unique_ptr<ParsedExpression> &expression) { result |= ContainsSimpleCase(*expression); });
+	return result;
+}
+
+static bool IsDynamicProjection(const ParsedExpression &expression) {
+	if (expression.GetExpressionClass() == ExpressionClass::STAR) {
+		return true;
+	}
+	for (auto &child : expression.Children()) {
+		if (IsDynamicProjection(child)) {
+			return true;
+		}
+	}
+	return false;
+}
+
+vector<unique_ptr<ParsedExpression>> SelectNode::GetSelectListForSerialization(Serializer &serializer) const {
+	vector<unique_ptr<ParsedExpression>> result;
+	result.reserve(select_list.size());
+	for (auto &expression : select_list) {
+		auto copy = expression->Copy();
+		if (!serializer.ShouldSerialize(StorageVersion::V2_0_0) && !expression->HasAlias() &&
+		    !IsDynamicProjection(*expression) && ContainsSimpleCase(*expression)) {
+			copy->SetAlias(expression->GetName());
+		}
+		result.push_back(std::move(copy));
+	}
+	return result;
 }
 
 string SelectNode::ToString() const {
